+++ src/app/routes/transactions/placeOrder.ts
@@ -48,7 +48,24 @@ placeOrderTransactionsRouter.post(
     validator,
     async (req, res, next) => {
         try {
+            auth.setCredentials({ access_token: req.accessToken });
+            const placeOrderService = new cinerinoapi.service.transaction.PlaceOrder4ttts({
+                auth: auth,
+                endpoint: <string>process.env.CINERINO_API_ENDPOINT
+            });
+            const sellerService = new cinerinoapi.service.Seller({
+                auth: auth,
+                endpoint: <string>process.env.CINERINO_API_ENDPOINT
+            });
+
             const sellerIdentifier = 'TokyoTower';
+            const searchSellersResult = await sellerService.search({
+                limit: 1
+            });
+            const seller = searchSellersResult.data.shift();
+            if (seller === undefined) {
+                throw new Error('Seller not found');
+            }
 
             // WAITER許可証を取得
             const scope = 'placeOrderTransaction.TokyoTower.POS';
@@ -64,16 +81,16 @@ placeOrderTransactionsRouter.post(
             const expires = moment(req.body.expires)
                 .toDate();
 
-            auth.setCredentials({ access_token: req.accessToken });
-            const placeOrderService = new cinerinoapi.service.transaction.PlaceOrder4ttts({
-                auth: auth,
-                endpoint: <string>process.env.CINERINO_API_ENDPOINT
-            });
-
             const transaction = await placeOrderService.start({
                 expires: expires,
                 sellerIdentifier: sellerIdentifier, // 電波塔さんの組織識別子(現時点で固定)
-                passportToken: token
+                passportToken: token,
+                ...{
+                    seller: {
+                        typeOf: seller.typeOf,
+                        id: seller.id
+                    }
+                }
             });
 
             // tslint:disable-next-line:no-string-literal