+++ CHANGELOG.md
@@ -18,6 +18,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v29.0.0 - 2019-10-31
+
+### Changed
+
+- POS注文取引確定時の決済方法指定を削除
+- 注文取引開始時の販売者をIDで指定するように調整
+- @cinerino/api-nodejs-client@2.0.0-alpha.35で再構築
+- POS使用サービスにposスコープを許可
+
 ## v28.1.1 - 2019-10-29
 
 ### Changed